Clean and fix tests. Add asserts if possible. Add fail where no assert.
