+++ tests/multilines.et
@@ -55,3 +55,4 @@ mstring in
 func args
 """)
 @#
+
